refactor: move auth feature's domain to domain module
