+++ feat/auth/src/main/java/com/emotionstorage/auth/presentation/SignupCompleteViewModel.kt
@@ -1,7 +1,7 @@
 package com.emotionstorage.auth.presentation
 
 import androidx.lifecycle.ViewModel
-import com.emotionstorage.auth.domain.usecase.LoginWithIdTokenUseCase
+import com.emotionstorage.domain.useCase.auth.LoginWithIdTokenUseCase
 import com.emotionstorage.domain.common.DataState
 import com.emotionstorage.domain.model.User
 import com.orhanobut.logger.Logger
