+++ feat/tutorial/build.gradle.kts
@@ -27,6 +27,5 @@ dependencies {
     implementation(projects.domain)
     implementation(projects.core.common)
     implementation(projects.core.ui)
-    implementation(projects.feat.auth)
     testImplementation(kotlin("test"))
 }
