+++ feat/tutorial/src/main/java/com/emotionstorage/tutorial/ui/onBoarding/GenderBirthScreen.kt
@@ -24,7 +24,7 @@ import androidx.compose.ui.tooling.preview.PreviewScreenSizes
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.hilt.navigation.compose.hiltViewModel
-import com.emotionstorage.auth.domain.model.SignupForm.GENDER
+import com.emotionstorage.domain.model.SignupForm.GENDER
 import com.emotionstorage.tutorial.R
 import com.emotionstorage.tutorial.presentation.onBoarding.GenderBirthEvent
 import com.emotionstorage.tutorial.presentation.onBoarding.GenderBirthViewModel
